+++ src/main/java/com/project/devgram/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.project.devgram.repository;
 
-import com.project.devgram.domain.comment.Comment;
+import com.project.devgram.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
